+++ flang/lib/Optimizer/Passes/Pipelines.cpp
@@ -325,8 +325,8 @@ void createDefaultFIRCodeGenPassPipeline(mlir::PassManager &pm,
 
   pm.addPass(fir::createFunctionAttr(
       {framePointerKind, config.NoInfsFPMath, config.NoNaNsFPMath,
-       config.ApproxFuncFPMath, config.NoSignedZerosFPMath,
-       config.UnsafeFPMath}));
+       config.ApproxFuncFPMath, config.NoSignedZerosFPMath, config.UnsafeFPMath,
+       ""}));
 
   fir::addFIRToLLVMPass(pm, config);
 }